refactor: ErrorAlertをcomponentsディレクトリへ移動
